Separate bssfp 1-pool and 2-pool
