+++ test/bssfp1.jl
@@ -1,5 +1,5 @@
 #=
-test/bssfp.jl
+test/bssfp1.jl 1-pool case
 =#
 
 using BlochSim: bssfp, Spin, InstantaneousRF, duration, real_imag
@@ -94,39 +94,4 @@ end
     tRF_ms = 1e-3 # super-short RF pulse
     sig7 = @inferred bssfp(bSSFPbloch3, tRF_ms, xt..., xs...)
     @test isapprox(sig4, sig7, atol = 2e-6)
-end
-
-
-# two-pool test
-@testset "bssfp2" begin
-# todo: currently all the values must be Float for jacobian to work
-
-    M0_phase = π/3
-    Mz0 = 0.9
-    f_f = 0.15 # fast fraction
-    T1_f_ms = 400. # T1 for fast-relaxing, myelin water compartment
-    T1_s_ms = 832. # T1 for slow-relaxing, non-myelin water compartment
-    T2_f_ms = 20. # T2 for fast-relaxing, myelin water compartment
-    T2_s_ms = 80. # T2 for slow-relaxing, non-myelin water compartment
-    τ_fs_ms = 50. # residence time
-    Δf_myelin_Hz = 5. # frequency shift of myelin water
-    Δf0_Hz = -40. # from B0
-    α_deg, TR_ms, TE_ms = 20, 10, 5 # scan parameters
-    α_rad = deg2rad(α_deg)
-    Δϕ_deg = 30 # RF phase cycling increment
-    Δϕ_rad = deg2rad(Δϕ_deg)
-
-    xt = (; M0_phase, Mz0, f_f, T1_f_ms, T1_s_ms, T2_f_ms, T2_s_ms,
-        τ_fs_ms, Δf_myelin_Hz, Δf0_Hz)
-    xs = (; Δϕ_rad, TR_ms, TE_ms, α_rad)
-    tmp1 = @inferred bssfp(xt..., xs...)
-
-    @test tmp1 isa Complex{<:AbstractFloat}
-
-    fun(xt) = real_imag(bssfp(xt..., xs...))
-    tmp2 = @inferred fun(xt)
-    @test real_imag(tmp1) == tmp2
-
-    grad = ForwardDiff.jacobian(fun, collect(xt))
-    @test grad isa Matrix{<:AbstractFloat}
 end
