+++ test/runtests.jl
@@ -17,7 +17,8 @@ end
     include("freeprecess.jl")
     include("excite.jl")
     include("rf-rect.jl")
-    include("bssfp.jl")
+    include("bssfp1.jl")
+    include("bssfp2.jl")
     include("mese.jl")
     include("spgr.jl")
     include("crb.jl")
